+++ cosmrs/src/lib.rs
@@ -1,17 +1,40 @@
 #![doc = include_str!("../README.md")]
 #![doc(
-    html_logo_url = "https://raw.githubusercontent.com/cosmos/cosmos-rust/main/.images/cosmos.png",
-    html_root_url = "https://docs.rs/cosmrs/0.5.1"
+    html_logo_url = "https://raw.githubusercontent.com/cosmos/cosmos-rust/main/.images/cosmos.png"
 )]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![forbid(unsafe_code)]
-#![warn(trivial_casts, trivial_numeric_casts, unused_import_braces)]
+#![warn(
+    clippy::checked_conversions,
+    clippy::panic,
+    clippy::panic_in_result_fn,
+    clippy::unwrap_used,
+    missing_docs,
+    trivial_casts,
+    trivial_numeric_casts,
+    rust_2018_idioms,
+    unused_lifetimes,
+    unused_import_braces
+)]
 
+//! ## Re-exports
+//!
+//! CosmRS re-exports the following crates for easy access:
+//!
+//! - `bip32`: re-exported as `cosmrs::bip32`
+//! - `cosmos-sdk-proto`: re-exported as `cosmrs::proto`
+//! - `tendermint`: re-exported as `cosmrs::tendermint`
+//! - `tendermint-rpc`: re-exported as `cosmrs::rpc` (requires `rpc` crate feature)
+
+pub mod abci;
+pub mod auth;
 pub mod bank;
 pub mod crypto;
 pub mod distribution;
+pub mod feegrant;
 pub mod staking;
 pub mod tx;
+pub mod vesting;
 
 #[cfg(feature = "cosmwasm")]
 pub mod cosmwasm;
@@ -21,20 +44,16 @@ pub mod cosmwasm;
 pub mod dev;
 
 mod base;
-mod decimal;
 mod error;
-mod prost_ext;
 
 pub use crate::{
-    base::{AccountId, Coin, Denom},
-    decimal::Decimal,
+    base::{AccountId, Amount, Coin, Denom, Gas},
     error::{Error, Result},
     tx::Tx,
 };
 
-pub use cosmos_sdk_proto as proto;
+pub use cosmos_sdk_proto::{self as proto, Any};
 pub use eyre::Report as ErrorReport;
-pub use prost_types::Any;
 pub use tendermint;
 
 #[cfg(feature = "bip32")]