+++ cosmrs/src/tx/body.rs
@@ -1,7 +1,9 @@
 //! Transaction bodies.
 
-use crate::{prost_ext::MessageExt, proto, ErrorReport, Result};
-use prost_types::Any;
+use crate::{
+    proto::{self, traits::MessageExt},
+    Any, ErrorReport, Result,
+};
 use tendermint::block;
 
 /// [`Body`] of a transaction that all signers sign over.
@@ -63,7 +65,13 @@ impl Body {
 
     /// Encode this type using Protocol Buffers.
     pub fn into_bytes(self) -> Result<Vec<u8>> {
-        self.into_proto().to_bytes()
+        Ok(self.into_proto().to_bytes()?)
+    }
+}
+
+impl Default for Body {
+    fn default() -> Body {
+        Self::new([], "", 0u8)
     }
 }