+++ cosmrs/src/tx/raw.rs
@@ -1,6 +1,12 @@
 //! Raw transaction.
 
-use crate::{prost_ext::MessageExt, proto, Result};
+use crate::{
+    proto::{
+        self,
+        traits::{Message, MessageExt},
+    },
+    Result,
+};
 
 #[cfg(feature = "rpc")]
 use crate::rpc;
@@ -16,12 +22,12 @@ pub struct Raw(proto::cosmos::tx::v1beta1::TxRaw);
 impl Raw {
     /// Deserialize raw transaction from serialized protobuf.
     pub fn from_bytes(bytes: &[u8]) -> Result<Self> {
-        Ok(Raw(prost::Message::decode(bytes)?))
+        Ok(Raw(Message::decode(bytes)?))
     }
 
     /// Serialize raw transaction as a byte vector.
     pub fn to_bytes(&self) -> Result<Vec<u8>> {
-        self.0.to_bytes()
+        Ok(self.0.to_bytes()?)
     }
 
     /// Broadcast this transaction using the provided RPC client
@@ -31,7 +37,7 @@ impl Raw {
     where
         C: rpc::Client + Send + Sync,
     {
-        Ok(client.broadcast_tx_commit(self.to_bytes()?.into()).await?)
+        Ok(client.broadcast_tx_commit(self.to_bytes()?).await?)
     }
 }